+++ DSCResources/MSFT_xExchOabVirtualDirectory/MSFT_xExchOabVirtualDirectory.psm1
@@ -39,7 +39,7 @@ function Get-TargetResource
         $ExtendedProtectionSPNList,
 
         [Parameter()]
-        [ValidateSet('None','Allow','Require')]
+        [ValidateSet('None', 'Allow', 'Require')]
         [System.String]
         $ExtendedProtectionTokenChecking,
 
@@ -50,6 +50,10 @@ function Get-TargetResource
         [Parameter()]
         [System.String]
         $InternalUrl,
+
+        [Parameter()]
+        [System.Boolean]
+        $OAuthAuthentication,
 
         [Parameter()]
         [System.Int32]
@@ -99,6 +103,7 @@ function Get-TargetResource
             ExternalUrl                     = [System.String] $vdir.ExternalUrl.AbsoluteUri
             InternalUrl                     = [System.String] $vdir.InternalUrl.AbsoluteUri
             OABsToDistribute                = [System.String[]] $oabNames
+            OAuthAuthentication             = [System.Boolean] $vdir.OAuthAuthentication
             PollInterval                    = [System.Int32] $vdir.PollInterval
             RequireSSL                      = [System.Boolean] $vdir.RequireSSL
             WindowsAuthentication           = [System.Boolean] $vdir.WindowsAuthentication
@@ -147,7 +152,7 @@ function Set-TargetResource
         $ExtendedProtectionSPNList,
 
         [Parameter()]
-        [ValidateSet('None','Allow','Require')]
+        [ValidateSet('None', 'Allow', 'Require')]
         [System.String]
         $ExtendedProtectionTokenChecking,
 
@@ -158,6 +163,10 @@ function Set-TargetResource
         [Parameter()]
         [System.String]
         $InternalUrl,
+
+        [Parameter()]
+        [System.Boolean]
+        $OAuthAuthentication,
 
         [Parameter()]
         [System.Int32]
@@ -253,7 +262,7 @@ function Test-TargetResource
         $ExtendedProtectionSPNList,
 
         [Parameter()]
-        [ValidateSet('None','Allow','Require')]
+        [ValidateSet('None', 'Allow', 'Require')]
         [System.String]
         $ExtendedProtectionTokenChecking,
 
@@ -264,6 +273,10 @@ function Test-TargetResource
         [Parameter()]
         [System.String]
         $InternalUrl,
+
+        [Parameter()]
+        [System.Boolean]
+        $OAuthAuthentication,
 
         [Parameter()]
         [System.Int32]
@@ -341,6 +354,11 @@ function Test-TargetResource
         {
             $testResults = $false
         }
+
+        if (!(Test-ExchangeSetting -Name 'OAuthAuthentication' -Type 'Boolean' -ExpectedValue $OAuthAuthentication -ActualValue $vdir.OAuthAuthentication -PSBoundParametersIn $PSBoundParameters -Verbose:$VerbosePreference))
+        {
+            $testResults = $false
+        }
     }
 
     return $testResults
@@ -386,7 +404,7 @@ function AddOabDistributionPoint
         $ExtendedProtectionSPNList,
 
         [Parameter()]
-        [ValidateSet('None','Allow','Require')]
+        [ValidateSet('None', 'Allow', 'Require')]
         [System.String]
         $ExtendedProtectionTokenChecking,
 
@@ -397,6 +415,10 @@ function AddOabDistributionPoint
         [Parameter()]
         [System.String]
         $InternalUrl,
+
+        [Parameter()]
+        [System.Boolean]
+        $OAuthAuthentication,
 
         [Parameter()]
         [System.Int32]